requirements.txt
tensorflow==2.6.0
matplotlib==3.4.3
scikit-learn==0.24.2
scipy==1.7.3
joblib==1.0.1
rdkit-pypi==2021.3.5
pandas==1.3.3
numpy==1.21.2